tests/wpt/tests/css/css-scoping/font-face-001.html

<!doctype html>
<title>CSS Test: @font-face applies in the shadow tree.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link rel="author" title="Alan Stearns" href="mailto:stearns@adobe.com">
<link rel="help" href="https://drafts.csswg.org/css-scoping/#shadow-names">

<style>
#in-document {
  font-family: ff-16-wide;
}
</style>
<div id="host"><span id="in-document">1234567890</span></div>
<script>
promise_test(async () => {
  host.attachShadow({ mode: "open" }).innerHTML = `
    <style>
      @font-face {
        font-family: ff-16-wide;
        src: url(/fonts/Ahem.ttf);
      }
      #in-shadow {
        font-family: ff-16-wide;
      }
    </style>
    <slot></slot>
    <span id="in-shadow">0123456789</span>
  `;

  await document.fonts.ready;

  assert_not_equals(document.getElementById('in-document').offsetWidth, 160);
  assert_equals(host.shadowRoot.getElementById('in-shadow').offsetWidth, 160);
}, "@font-face applies in the shadow tree")
</script>
