third_party/blink/web_tests/external/wpt/css/css-scoping/font-face-001-expected.txt

This is a testharness.js-based test.
[FAIL] @font-face applies in the shadow tree
  assert_equals: expected 160 but got 80
Harness: the test ran to completion.
